@@ -1,38 +1,39 @@
 {
   "name": "sarangjo.github.io",
   "private": true,
-  "description": "A starter for a blog powered by Gatsby and Markdown",
+  "description": "Sarang Joshi's personal website",
   "version": "0.1.0",
-  "author": "Kyle Mathews <mathews.kyle@gmail.com>",
+  "author": "Sarang Joshi <sarangj@msn.com>",
   "bugs": {
     "url": "https://github.com/gatsbyjs/gatsby/issues"
   },
   "dependencies": {
-    "@material-ui/core": "^4.8.3",
-    "gatsby": "^3.3.0",
-    "gatsby-image": "^3.3.0",
-    "gatsby-plugin-google-analytics": "^3.3.0",
-    "gatsby-plugin-manifest": "^3.3.0",
-    "gatsby-plugin-offline": "^4.3.0",
-    "gatsby-plugin-react-helmet": "^4.3.0",
-    "gatsby-plugin-sharp": "^3.3.0",
-    "gatsby-plugin-typography": "^3.3.0",
-    "gatsby-remark-autolink-headers": "^4.0.0",
-    "gatsby-remark-copy-linked-files": "^4.0.0",
+    "@emotion/react": "^11.7.1",
+    "@emotion/styled": "^11.6.0",
+    "@mui/material": "^5.2.8",
+    "gatsby": "^4.5.2",
+    "gatsby-plugin-google-analytics": "^4.5.0",
+    "gatsby-plugin-image": "^2.5.1",
+    "gatsby-plugin-manifest": "^4.5.1",
+    "gatsby-plugin-offline": "^5.5.1",
+    "gatsby-plugin-react-helmet": "^5.5.0",
+    "gatsby-plugin-sharp": "^4.5.1",
+    "gatsby-plugin-typography": "^4.5.0",
+    "gatsby-remark-autolink-headers": "^5.5.0",
+    "gatsby-remark-copy-linked-files": "^5.5.0",
     "gatsby-remark-external-links": "^0.0.4",
-    "gatsby-remark-images": "^5.0.0",
-    "gatsby-remark-prismjs": "^5.0.0",
+    "gatsby-remark-images": "^6.5.1",
+    "gatsby-remark-prismjs": "^6.5.0",
     "gatsby-remark-reading-time": "^1.1.0",
-    "gatsby-remark-responsive-iframe": "^4.0.0",
-    "gatsby-remark-smartypants": "^4.0.0",
-    "gatsby-source-filesystem": "^3.3.0",
-    "gatsby-transformer-remark": "^4.0.0",
-    "gatsby-transformer-sharp": "^3.3.0",
-    "lodash": "^4.17.15",
+    "gatsby-remark-responsive-iframe": "^5.5.0",
+    "gatsby-remark-smartypants": "^5.5.0",
+    "gatsby-source-filesystem": "^4.5.1",
+    "gatsby-transformer-remark": "^5.5.1",
+    "gatsby-transformer-sharp": "^4.5.0",
     "prismjs": "^1.15.0",
-    "react": "^16.8.0",
-    "react-dom": "^16.8.0",
-    "react-helmet": "^5.2.0",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "react-helmet": "^6.1.0",
     "react-typography": "^0.16.18",
     "typeface-merriweather": "0.0.54",
     "typeface-montserrat": "0.0.54",
@@ -55,12 +56,16 @@
     "start": "yarn develop",
     "serve": "gatsby serve",
     "predeploy": "gatsby build",
-    "deploy": "gh-pages -d public -b master",
+    "deploy": "gh-pages -d public -b deploy",
     "test": "echo \"Write tests! -> https://gatsby.app/unit-testing\""
   },
   "devDependencies": {
     "eslint": "5.16.0",
     "eslint-plugin-react": "^7.16.0",
     "gh-pages": "^2.0.1"
+  },
+  "engines": {
+    "node": "14.x",
+    "yarn": "^1.15.0"
   }
 }
